Update travel distances migration to support bigint IDs and add corresponding tests
Modified the travel distances migration functions to utilize bigint data types for from_id and to_id, accommodating larger IDs from production. Added a new test suite to verify the correct handling of bigint IDs in the migration process, ensuring that SQL queries reflect these changes. This update enhances compatibility with larger datasets and improves the robustness of the migration logic.

+++ distance_migration/travel_distances_migration.py
@@ -195,7 +195,7 @@ def get_existing_pairs_for_keys(connection, from_type, to_type, method, pair_key
                 FROM travel_distances
                 WHERE from_type = %s AND to_type = %s AND travel_method = %s
                   AND (from_id, to_id) IN (
-                    SELECT * FROM UNNEST(%s::int[], %s::int[]) AS t(from_id, to_id)
+                    SELECT * FROM UNNEST(%s::bigint[], %s::bigint[]) AS t(from_id, to_id)
                   )
             """, (from_type, to_type, method, from_ids, to_ids))
             for row in cursor.fetchall():
@@ -324,9 +324,9 @@ def _copy_insert_batch(cursor, rows, skip_conflict_check=False):
     cursor.execute("""
         CREATE TEMP TABLE IF NOT EXISTS _td_stage (
             from_type text,
-            from_id int,
+            from_id bigint,
             to_type text,
-            to_id int,
+            to_id bigint,
             travel_method text,
             distance_meters int,
             duration_minutes int,
